requirements.txt
fastapi
   uvicorn
   httpx
   python-dotenv
   fastapi[standard]
   pydantic-settings
   crawl4ai @ git+https://github.com/unclecode/crawl4ai.git